Fix round prototypes for windows

+++ platform.h
@@ -58,7 +58,7 @@ extern "C" {
 #elif defined(_WIN32) || defined(_WIN64) 
     /* does not exist on windows */
     char * strndup( const char *s1, size_t n);
-    float round (float v);
+    double round (float v);
 #    pragma warning (disable: 4244) // suspend warnings
 #endif // _WIN32 || _WIN64
